영농일지 작성자 수정/삭제 버튼 안보이는 오류 수정 마이페이지 alert창 수정 follow 기능 com.kh.tinyfamr/follw 쪽으로 전체 이동
