+++ tinyFarm/src/main/java/com/kh/tinyfarm/diary/controller/MypageController.java
@@ -339,114 +339,6 @@ public class MypageController {
 		return result;
 	}
 	
-	//팔로잉 모달창 정보
-	@ResponseBody
-	@PostMapping(value = "getFollowingInfo.me", produces = "application/json; charset=UTF-8")
-	public Member getFollowingInfo(String followingId) {
-		Member m = diaryService.selectFollowingInfo(followingId); //팔로우 한 유저 아이디 정보 가져오기
-		return m;
-	}
-	//팔로워 모달창 정보
-	@ResponseBody
-	@PostMapping(value = "getFollowerInfo.me", produces = "application/json; charset=UTF-8")
-	public Member getFollowerInfo(String followerId) {
-		Member m = diaryService.selectFollowerInfo(followerId); //팔로우 한 유저 아이디 정보 가져오기
-		return m;
-	}
-	//팔로우 확인
-	@ResponseBody
-	@RequestMapping("followChk.me")
-	public String followCheck(int userNo,String followingId) {
-		
-		Follow f = new Follow();
-		f.setUserNo(userNo);
-		f.setFollowingId(followingId);
-		
-		int result = diaryService.followCheck(f);
-		
-		if(result==1) {
-			return "YY";
-		}else {
-			return "NN";
-		}
-	}
-	//팔로우
-	@RequestMapping("follow.me")
-	public String followUser(String followingId,HttpSession session, HttpServletRequest request) {
-		//로그인 유저 정보
-		Member loginUser = (Member)session.getAttribute("loginUser");
-		int userNo = loginUser.getUserNo();
-		
-		Follow f = new Follow();
-		f.setUserNo(userNo);
-		f.setFollowingId(followingId);
-		
-		//이전페이지
-		String before = (String)request.getHeader("Referer");
-		
-		int result = diaryService.followUser(f);
-		
-		if(result>0) {	
-			session.setAttribute("alertMsg", "팔로우 완료!");
-		}else {
-			session.setAttribute("alertMsg", "팔로우 실패. 다시 확인해주세요.");
-		}
-		//이전페이지(보고있던 페이지)로 리턴
-		return "redirect:"+before;
-		
-	}
-	
-	//언팔로우
-	@PostMapping("unfollow.me")
-	public String unfollowUser(String followingId,HttpSession session, HttpServletRequest request) {
-		
-		//로그인 유저 정보
-		Member loginUser = (Member)session.getAttribute("loginUser");
-		int userNo = loginUser.getUserNo();
-		
-		Follow f = new Follow();
-		f.setUserNo(userNo);
-		f.setFollowingId(followingId);
-		
-		//이전페이지
-		String before = (String)request.getHeader("Referer");
-		
-		int result = diaryService.unfollowUser(f);
-		//언팔 결과 확인 후 활동내역 페이지 유지
-		if(result>0) {
-			session.setAttribute("alertMsg", "팔로우 취소가 완료되었습니다.");
-		}else {
-			session.setAttribute("alertMsg", "다시 시도해주세요.");
-		}
-		//이전페이지(보고있던 페이지)로 리턴
-		return "redirect:"+before;
-	}
-	
-	//존재하지 않는 회원 목록에서 삭제
-	@PostMapping("deleteNonUser.me")
-	public String deleteNonUser(String followingId,HttpSession session, HttpServletRequest request) {
-		
-		//로그인 유저 회원정보
-		Member loginUser = (Member)session.getAttribute("loginUser");
-		int userNo = loginUser.getUserNo();
-		
-		Follow f = new Follow();
-		f.setUserNo(userNo);
-		f.setFollowingId(followingId);
-		int result = diaryService.unfollowUser(f);
-		
-		if(result>0) {
-			session.setAttribute("alertMsg", "삭제 완료");
-		}else {
-			session.setAttribute("alertMsg", "다시 시도해주세요.");
-		}
-		
-		//이전페이지
-		String before = (String)request.getHeader("Referer");
-		
-		//이전페이지(보고있던 페이지)로 리턴
-		return "redirect:"+before;
-	}
 	//유저 영농일지 보기
 	@PostMapping("follow.di")
 	public String followDiaryView(String followingId, Model model) {
@@ -728,7 +620,6 @@ public class MypageController {
 		
 		//좋아요 정보 가져오기
 		DiaryLike like = diaryService.selectLike(dl);
-		System.out.println(like);
 		
 		//좋아요 수
 		int likeCount = diaryService.countLike(diaryNo);
@@ -904,11 +795,6 @@ public class MypageController {
 		ArrayList<Product> wishList = diaryService.selectWish(userNo);
 		result.put("wishList", wishList);
 		return result;
-	}
-	
-	@GetMapping("modal.me")
-	public String loadModal() {
-		return "member/memberModal";
 	}
