+++ tinyFarm/src/main/java/com/kh/tinyfarm/diary/model/service/DiaryServiceImpl.java
@@ -186,28 +186,6 @@ public class DiaryServiceImpl implements DiaryService {
 	public ArrayList<Product> mySellList(int userNo, PageInfo sPi) {
 		return (ArrayList)diaryDao.mySellList(sqlSession,userNo,sPi);
 	}
-	//팔로잉 정보
-	@Override
-	public Member selectFollowingInfo(String followingId) {
-		return diaryDao.selectFollowInfo(sqlSession,followingId);
-	}
-	//팔로워 정보
-	@Override
-	public Member selectFollowerInfo(String followingId) {
-		return diaryDao.selectFollowInfo(sqlSession,followingId);
-	}
-	
-	//팔로우 걸기
-	@Override
-	public int followUser(Follow f) {
-		return diaryDao.followUser(sqlSession,f);
-	}
-
-	//팔로우 취소
-	@Override
-	public int unfollowUser(Follow f) {
-		return diaryDao.unfollowUser(sqlSession,f);
-	}
 	
 	//카카오 로그인을 위한 유저 정보
 	@Override
@@ -244,17 +222,6 @@ public class DiaryServiceImpl implements DiaryService {
 	public ArrayList<Product> selectWish(int userNo) {
 		return (ArrayList)diaryDao.selectWish(sqlSession,userNo);
 	}
-	//팔로우 여부 확인
-	@Override
-	public int followCheck(Follow f) {
-		return diaryDao.followCheck(sqlSession,f);
-	}
-	
-	//상태 Y인 유저 정보
-	@Override
-	public Member selectFollowMember(String statusYFiId) {
-		return diaryDao.selectFollowMember(sqlSession,statusYFiId);
-	}
 	
 	//좋아요 정보
 	@Override
@@ -285,8 +252,5 @@ public class DiaryServiceImpl implements DiaryService {
 	public Member selectDiaryWriter(int userNo) {
 		return diaryDao.selectDiaryWriter(sqlSession, userNo);
 	}
-	
-	
-	
 	
 }
