+++ tinyFarm/src/main/java/com/kh/tinyfarm/follow/model/service/FollowServiceImpl.java
@@ -7,6 +7,7 @@ import org.springframework.stereotype.Service;
 
 import com.kh.tinyfarm.follow.model.dao.FollowDao;
 import com.kh.tinyfarm.follow.model.vo.Follow;
+import com.kh.tinyfarm.member.model.vo.Member;
 
 @Service
 public class FollowServiceImpl implements FollowService{
@@ -36,5 +37,23 @@ public class FollowServiceImpl implements FollowService{
 		
 		return followDao.selectFollow(sqlSession,fw);
 	}
+	
+	//팔로우 여부 확인
+	@Override
+	public int followCheck(Follow f) {
+		return followDao.followCheck(sqlSession,f);
+	}
+	
+	//팔로잉 정보
+	@Override
+	public Member selectFollowingInfo(String followingId) {
+		return followDao.selectFollowInfo(sqlSession,followingId);
+	}
+	//팔로워 정보
+	@Override
+	public Member selectFollowerInfo(String followingId) {
+		return followDao.selectFollowInfo(sqlSession,followingId);
+	}
+
 
 }
